Additional renamings as aligned with TK.
