+++ test/aws_mock_data.py
@@ -12,8 +12,8 @@ from exasol.ds.sandbox.lib.aws_access.snapshot import Snapshot
 from exasol.ds.sandbox.lib.aws_access.stack_resource import StackResource
 from test.conftest import DEFAULT_ASSET_ID
 
-TEST_ROLE_ID = 'VM-SLC-Bucket-VMImportRole-TEST'
-TEST_BUCKET_ID = 'vm-slc-bucket-vmslcbucket-TEST'
+TEST_ROLE_ID = 'VM-DSS-Bucket-VMImportRole-TEST'
+TEST_BUCKET_ID = 'vm-dss-bucket-vmdssbucket-TEST'
 TEST_AMI_ID = "AMI-IMAGE-12345"
 TEST_CLOUDFRONT_ID = "test-cloudfrontet-TEST"
 TEST_CLOUDFRONT_DOMAIN_NAME = "test-s3.cloudfront.net"
@@ -43,7 +43,7 @@ def get_ami_image_mock_data(state: str) -> Ami:
                 'Name': DEFAULT_ASSET_ID.ami_name,
                 'RootDeviceName': '/dev/sda1',
                 'RootDeviceType': 'ebs', 'SriovNetSupport': 'simple',
-                'Tags': [{'Key': 'exa_slc_id', 'Value': DEFAULT_ASSET_ID.tag_value}],
+                'Tags': [{'Key': 'exa_dss_id', 'Value': DEFAULT_ASSET_ID.tag_value}],
                 'VirtualizationType': 'hvm'})
 
 
@@ -53,7 +53,7 @@ def get_snapshot_mock_data():
         'OwnerId': '123', 'Progress': '100%', 'SnapshotId': 'snap-123',
         'StartTime': datetime.datetime(2022, 8, 16, 15, 3, 40, 662000, tzinfo=tzutc()), 'State': 'completed',
         'VolumeId': 'vol-123', 'VolumeSize': 100,
-        'Tags': [{'Key': 'exa_slc_id', 'Value': DEFAULT_ASSET_ID.tag_value}],
+        'Tags': [{'Key': 'exa_dss_id', 'Value': DEFAULT_ASSET_ID.tag_value}],
         'StorageTier': 'standard'
     })
 
@@ -71,7 +71,7 @@ def get_export_image_task_mock_data(in_progress: bool):
             'Progress': "50%",
             "StatusMessage": "creating the image",
             'Status': 'active',
-            'Tags': [{'Key': 'exa_slc_id', 'Value': DEFAULT_ASSET_ID.tag_value}]
+            'Tags': [{'Key': 'exa_dss_id', 'Value': DEFAULT_ASSET_ID.tag_value}]
         })
     else:
         return ExportImageTask({
@@ -83,7 +83,7 @@ def get_export_image_task_mock_data(in_progress: bool):
                     'S3Prefix': DEFAULT_ASSET_ID.bucket_prefix
                 },
             'Status': 'completed',
-            'Tags': [{'Key': 'exa_slc_id', 'Value': DEFAULT_ASSET_ID.tag_value}]
+            'Tags': [{'Key': 'exa_dss_id', 'Value': DEFAULT_ASSET_ID.tag_value}]
         })
 
 
@@ -107,7 +107,7 @@ def get_ec2_cloudformation_mock_data():
         'RollbackConfiguration': {},
         'StackStatus': 'CREATE_COMPLETE',
         'DisableRollback': False, 'NotificationARNs': [], 'Capabilities': ['CAPABILITY_IAM'],
-        'Tags': [{'Key': 'exa_slc_id', 'Value': DEFAULT_ASSET_ID.tag_value}],
+        'Tags': [{'Key': 'exa_dss_id', 'Value': DEFAULT_ASSET_ID.tag_value}],
         'DriftInformation': {'StackDriftStatus': 'NOT_CHECKED'}
     })
 
@@ -137,7 +137,7 @@ def get_ec2_key_pair_mock_data():
         'KeyFingerprint': '12:34:56:78:90:12:34:56:78:90:12:34:56:78:90:12:34:56:78:90',
         'KeyName': 'ec2-key-test-key',
         'KeyType': 'rsa',
-        'Tags': [{'Key': 'exa_slc_id', 'Value': DEFAULT_ASSET_ID.tag_value}],
+        'Tags': [{'Key': 'exa_dss_id', 'Value': DEFAULT_ASSET_ID.tag_value}],
         'CreateTime': datetime.datetime(2022, 8, 16, 15, 30, 41,
                                         tzinfo=tzutc())
     })
@@ -146,14 +146,14 @@ def get_ec2_key_pair_mock_data():
 def get_s3_cloudformation_mock_data() -> List[CloudformationStack]:
     return [CloudformationStack({
         'StackId': 'test-s3-stack-id',
-        'StackName': "DATA-SCIENCE-SANDBOX-VM-SLC-Bucket",
+        'StackName': "DATA-SCIENCE-SANDBOX-VM-Bucket",
         'ChangeSetId': 'test-stack-changeset-id-2',
         'CreationTime': datetime.datetime(2022, 8, 16, 14, 30, 45, 559000, tzinfo=tzutc()),
         'LastUpdatedTime': datetime.datetime(2022, 8, 16, 14, 30, 51, 667000, tzinfo=tzutc()),
         'RollbackConfiguration': {},
         'StackStatus': 'CREATE_COMPLETE',
         'DisableRollback': False, 'NotificationARNs': [], 'Capabilities': ['CAPABILITY_IAM'],
-        'Tags': [{'Key': 'exa_slc_id', 'Value': DEFAULT_ASSET_ID.tag_value}],
+        'Tags': [{'Key': 'exa_dss_id', 'Value': DEFAULT_ASSET_ID.tag_value}],
         'DriftInformation': {'StackDriftStatus': 'NOT_CHECKED'},
         'Outputs': [{'OutputKey': 'VMBucketId',
                      'OutputValue': TEST_BUCKET_ID, 'Description': ''},
@@ -171,14 +171,14 @@ def get_s3_cloudformation_mock_data() -> List[CloudformationStack]:
 def get_waf_cloudformation_mock_data() -> List[CloudformationStack]:
     return [CloudformationStack({
         'StackId': 'test-waf-stack-id',
-        'StackName': "DATA-SCIENCE-SANDBOX-VM-SLC-Bucket-WAF",
+        'StackName': "DATA-SCIENCE-SANDBOX-VM-Bucket-WAF",
         'ChangeSetId': 'test-stack-changeset-id-3',
         'CreationTime': datetime.datetime(2022, 8, 16, 14, 30, 45, 559000, tzinfo=tzutc()),
         'LastUpdatedTime': datetime.datetime(2022, 8, 16, 14, 30, 51, 667000, tzinfo=tzutc()),
         'RollbackConfiguration': {},
         'StackStatus': 'CREATE_COMPLETE',
         'DisableRollback': False, 'NotificationARNs': [], 'Capabilities': [],
-        'Tags': [{'Key': 'exa_slc_id', 'Value': DEFAULT_ASSET_ID.tag_value}],
+        'Tags': [{'Key': 'exa_dss_id', 'Value': DEFAULT_ASSET_ID.tag_value}],
         'DriftInformation': {'StackDriftStatus': 'NOT_CHECKED'},
         'Outputs': [{'OutputKey': 'VMDownloadACLArn',
                      'OutputValue': TEST_ACL_ARN, 'Description': ''}
