+++ test/cloudformation_validation.py
@@ -6,7 +6,7 @@ def validate_using_cfn_lint(tmp_path, cloudformation_yml):
     This test uses cfn-lint to validate the Cloudformation template.
     (See https://github.com/aws-cloudformation/cfn-lint)
     """
-    out_file = tmp_path / "slc_source_credential.yaml"
+    out_file = tmp_path / "cloudformation.yaml"
     with open(out_file, "w") as f:
         f.write(cloudformation_yml)
